fix(#92): Resolve Clean Code violations in /scripts folder
## Changes

### 1. Function Naming (CC-2)
- Rename `submit_log_running_task` → `submit_long_running_task` for clarity

### 2. DRY: Extract Duplicate Ollama Helpers
- Create `scripts/_ollama.py` with shared `_post_json`, `call_ollama_blocking`,
  and `call_ollama_stream` functions
- Update `eval_models.py` and `query_model.py` to import from `_ollama`
- Update `rag_query.py` to import shared utilities while keeping RAG-specific
  wrappers

### 3. Magic Numbers (CC-4)
- Add `HTTP_CONFLICT = 409` constant in `seed_demo.py`
- Replace hardcoded 409 comparison with named constant
- Remove PLR2004 suppression

### 4. Module-Level Imports (PEP 8)
- Move `import uuid` from function body to module-level in `seed_demo.py`

### 5. Encapsulation
- Add public `authorization_header` property to `sdk/gideon/client.py`
- Update `upload_file.py` to use public property instead of private attribute
- Remove SLF001 suppression from `upload_file.py`

All violations from issue #92 resolved. Code quality improved through
refactoring without introducing new issues.
